Move Compiler class into its own file.
